Work around AWS's overzealous URI encoding demands.
`+` is an allowable character in a URI's path. It is within the `sub-delims`
class defined by rfc3986, and path characters explictly include
`sub-delims` as allowable. AWS demands that it be encoded in the
canonical public URIs for objects containing `+` in their name, as we
must do (since the object names include semantic version data, and `+`
is the designated delimiter for post-version-number information).

Rather than argue the case, it's easier just to tell Addressable to
escape the `+` in the normalized form of the URI.