Set a configurable default user for package paths.
A package path now consists of an OPTIONAL user, mandatory package name,
and optional version. If the username is omitted, the `default_user`
configured in `/opt/clusterware/etc/forge/config.yml` is used; _that_
value defaults to `alces` if unspecified.

We also now allow explicitly stating `latest` as a version.

Note that there's potential ambiguity if only two of the three parts of
a package path are specified: is it user/package or package/version? We
try to parse the second part as a version or `latest`, and if
successful, treat it as a version; if unsuccessful, treat it as a
package name and the first part as a user.