Assume the install failed if anything exits non 0
For the most part, the install scripts run a bunch of commands that
should all pass. As this is the more typical behaviour, the file is
sourced with `set -e`.

If a file does expect to return a non-zero status, then it can be
manually turned off with `set +e`