Allow `forge` to be configured by the flight config
FlightDirect has a handy `config set` command that allows for persistent
configs to be set. However it is opinionated and will only set env vars
that start with `FL_CONFIG`.

As such, `forge` has been updated to recognise when its `api` and `sso`
urls have been set in this manner. The new `FL_CONFIG` values take the
highest priority.

The version number has been bumped for the next release