Do not store the version in the env var
Storing the version in the env var makes it hard to determine the
rails public directory (as it is off the base URL).

It can be assumed that the version will only ever be bumped in the event
of a breaking change. In this case, the `forge-cli` is already tied to
the old version and will need to be updated anyway.

To make this transition easier, the env var has been renamed the
`CACHE_URL`. Thus it is clear that the env var is used for more than
just forge

The `SSO_URL` env var has been removed as it is now inconsistent with
the cache. This should only been required by the production server for
the time being. A way to set it (based on env var) can be added in the
future