Simplify the logic for setting the `api_url`
Previously the behaviour was highly inconsistent on whether the
`API_VERSION` will be appended `api_url`.

A breaking change has already been made when the default url was updated
to include `v1`. Instead all the base urls should be appended with the
version. This makes setting the url internally consistent.