Extract the method for setting the DATA_ROOT
Independent to `forge` ruby `LOAD_PATH` is the root location for its
configs and user data. Ideally installing `forge` would be supported in
both a `flight` and `clusterware` environment.

As such, the setting this `DATA_ROOT` has been extracted and will
default to `FL_ROOT`. It will then fall back to `cw_ROOT`.