Source the config before running the installer
Before `install.sh` is ran, the config is sourced. As the config is
optional, this step is skipped if missing.

The purpose of the install config is to set environment variables before
installing. The `set -a` flag causes all variables to be exported
automatically. It thus `export` does not need to be explicitly called in
the configs