Format the error on a new line for heredocs
The `shell` command always implicitly `cd` to a directory before running
the command. As such it prepends the `cd` command to the error message
to make it clear to the user.

Previously it would do this with ' && ' as the separator, however this
looks odd for cmds with multiple lines. Instead `heredoc` commands
should be prepended with a newline