Allow the package version to change in the Registry
Now that specific version of packages can be installed, it becomes
possible to reinstall a different version. Whilst this isn't
recommended, `forge` should handle the Registry correctly.

This simplest fix is to remove the previous Registry entry and then add
the new one. The new one will include the updated version number